requirements.txt
# =============================================================================
# Extracción de datos
# =============================================================================
requests>=2.31.0
python-dotenv>=1.0.0

# =============================================================================
# Manipulación de datos
# =============================================================================
pandas>=2.0.3
numpy>=1.24.0

# =============================================================================
# Exportación a Excel (opcional)
# =============================================================================
openpyxl>=3.1.2

# =============================================================================
# Comparación de productos entre supermercados
# =============================================================================
rapidfuzz>=3.5.0

# =============================================================================
# Dashboard e interfaz web
# =============================================================================
streamlit>=1.29.0
plotly>=5.18.0

# =============================================================================
# Automatización de cookies (opcional)
# =============================================================================
# selenium>=4.15.0
# playwright>=1.40.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
